# Conflicts:
#	src/main/java/com/webenglish/webenglish/Common/Role.java

+++ src/main/java/com/webenglish/webenglish/Model/Exams.java
@@ -14,11 +14,10 @@ import java.util.List;
 @Entity
 @Table(name = "EXAMS")
 public class Exams {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "EXAMID")
-    private Integer id;
+    private long id;
 
     @Column(name = "EXAMNAME", length = 100)
     private String examName;
